refactor(lotto): 상수 클래스 통한 상수 관리
