+++ src/main/java/constants/ConstantNumbersManager.java
@@ -1,0 +1,12 @@
+package constants;
+
+public class ConstantNumbersManager {
+    public static final int LOTTO_PRICE = 1000;
+    public static final int INDEX_RESET = 0;
+    public static final int LOTTO_SIZE = 6;
+    public static final int FIRST_INDEX = 1;
+    public static final int SECOND_INDEX = 2;
+    public static final int THIRD_INDEX = 3;
+    public static final int FORTH_INDEX = 4;
+    public static final int FIFTH_INDEX = 5;
+}
